+++ packages/eyes-sdk-core/lib/config/AccessibilityGuidelinesVersion.js
@@ -1,14 +1,16 @@
 const Enum = require('../utils/Enum')
 
 /**
- * @typedef {string} AccessibilityGuidelinesVersion
+ * @typedef {import('../utils/Enum').EnumValues<typeof AccessibilityGuidelinesVersions>} AccessibilityGuidelinesVersion
  */
 
 /**
  * The spec version for which to check the image visual accessibility level.
  */
 const AccessibilityGuidelinesVersions = Enum('AccessibilityGuidelinesVersion', {
+  /** @type {'WCAG_2_0'} */
   WCAG_2_0: 'WCAG_2_0',
+  /** @type {'WCAG_2_1'} */
   WCAG_2_1: 'WCAG_2_1',
 })