+++ packages/eyes-sdk-core/lib/config/AccessibilityLevel.js
@@ -1,16 +1,22 @@
 const Enum = require('../utils/Enum')
 
 /**
- * @typedef {string} AccessibilityLevel
+ * @typedef {import('../utils/Enum').EnumValues<typeof AccessibilityLevels>} AccessibilityLevel
  */
 
 /**
  * The extent in which to check the image visual accessibility level.
  */
 const AccessibilityLevels = Enum('AccessibilityLevel', {
-  /** Low accessibility level. */
+  /**
+   * Low accessibility level.
+   * @type {'AA'}
+   */
   AA: 'AA',
-  /** Highest accessibility level. */
+  /**
+   * Highest accessibility level.
+   * @type {'AAA'}
+   */
   AAA: 'AAA',
 })