+++ packages/sdk-coverage-tests/src/index.js
@@ -1,4 +1,3 @@
 const {makeEmitTracker} = require('./code-export')
-const coverageTests = require('./tests')
 
-module.exports = {makeEmitTracker, coverageTests}
+module.exports = {makeEmitTracker}