+++ packages/dom-snapshot/CHANGELOG.md
@@ -4,6 +4,10 @@
 ## Unreleased
 
 
+## 3.7.2 - 2020/7/22
+
+- avoid removing single `?` from query param in resource urls (mainly to address `?#iefix`, see https://github.com/applitools/mono/issues/79)
+
 ## 3.7.1 - 2020/7/16
 
 - fix selected options in select element ([Trello](https://trello.com/c/PftuuFqx))