+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- updated enum types to be more explicit about values
+
+## 11.3.9 - 2020/7/22
+
+- avoid `error.getTestResults is not a function` error ([Trello 1](https://trello.com/c/XAmA255U) [Trello 2](https://trello.com/c/1Bl2EaDE))
 
 ## 11.3.8 - 2020/7/19