+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 
+## 14.5.13 - 2020/7/22
+
+- updated to @applitools/dom-snapshot@3.7.2 (from 3.7.1)
+- updated to @applitools/eyes-sdk-core@11.3.9 (from 11.3.8)
+
 ## 14.5.12 - 2020/7/19
 
 - updated to @applitools/eyes-sdk-core@11.3.8 (from 11.3.7)